Add get_task to Client
Why:

When a task is being executed, there may be a need to access attributes
of the task such as tags. Since the client is already available,
otherwise the operator could not have been registered from there, a
simple way to provide access to the current task is through the client
object.

How:

The Dispatcher class saves a weakref to each of its instances. When
picking a task for execution, dispatchers keep a ref to the picked task.
Then, the client can import Dispatcher and get access to the instances
and to the picked task that is referenced.