Sort attribute's history
Why:

The events are not necessarily fetched in order from the dataset.
Setting the history in a random order would be terrible since the order
matters to execute the replay. This may have led to bugs where multiple
workers were able to reserve a task, leading to an history with
several reserved status in row.