Improve schedule efficiency
Why:

Maintaining tasks is too expensive for a process that runs on the login
nodes of clusters. The decision of scheduling should be taken on not
up-to-date reports, period. If there is enough maintainers being
executed, not having the scheduler to maintain prior to query should not
make much difference.