Change structure of event objects
Why:

The creation_timestamp is not reliable since there is a delay between
the timestamp and the actual registration in the database. We should
rather use the database backend mechanisms to create timestamps
automatically. For MongoDB, it is the '_id' which has an attribute
`generation_time`.

How:

Get rid of automatic id and rather set it as a unique field `key`. The
race conditions for event registrations are detected through this unique
key. The backend database takes care of converting its own `id` (`_id`
for mongodb) to `id`. Note that right now this is assumed to be an
ObjectId but this would not scale to other backends than MongoDB, since
the timestamp is dependent on the value of the id itself. We can't
convert random ids to ObjectId and assign a timestamp.