Add heartbeat and lost-task maintainer
Why:

When tasks are lost during execution, their status stays to Running.
Using a heartbeat we cannot infer if the task is lost and turn it to
FailedOver.

How:

Heartbeat is configurable. Since it can change from one task to another,
the current value is saved in every new task. When trying to infer if
the task is lost, we use the heartbeat of the task, not the current
global configuration.