Reduce race-conditions by catching some up-front
Why:

The reservation of tasks is the most common source of race conditions.
When there is a race-condition that is broad enough, the second request
passes to the update_status while status has changed to 'Reserved'.
Since 'Reserved' is a status with a potential heartbeat
(we don't want lost tasks with status 'Reserved') the second request is
valid and we end up with a task reserved by two different workers.

How:

By default, update_status will use the most recent status in the history
without refreshing it. This is to make sure we use the status the agent
making the request was aware of. If however the request is `reserve()`,
then we refresh the status and make sure it is specifically `Queued`
before passing it as-is to `update_status()`. This way, there can be no
race-conditions that are badly handled leading to a double reservation.