Use a queue to register live std{out,err}
Why:

It is tedious to do the registration live inside a forked process
because we need to re-instantiate a registrar inside it (MongoDB client
are not thread-safe). It is simpler to queue calls to std{out,err}
inside the forked process and empty the queue live inside the main
process. Main process is sleeping until the task is done, so better do
this than sleeping.