Fix reports of new output
Why:

The first report was not containing any output since it was a report for
new task. On completion, the report was only updating the event source
attributes. Since output and volume are not part of them, they were not
updated.

How:

Add an argument `update_output` to the method `update_report`. This
argument should only be used in `worker` at the completion point of the
task.