Drop maintainers from worker
Why:

That many maintainers inside each worker is too much. With a 100 workers
we end up with 100 * 5 maintainers, creating 500 connections to the
database. That is a crazy amount. Furthermore, 1 or 2 maintainers would
be well enough for a 100 workers. We should add a maintainer command
line, and maybe add automated scheduler of maintainers along with
workers.