Remove registrar attr from dispatcher
Why:

The dispatchers are referenced in the Class.__refs__ object. When adding
a reference to the registrar, we end up copying the registrar in the
forked processes. This is undesirable since registrar is not
thread-safe, and may uselessly lead to increased number of DB connections.

How:

Pass registrar as an argument to `pick()`.