Sort tasks based on `reported_on`
Why:

Worker should try to pick oldest tasks. If a task is interrupted if
should be accounted as old, otherwise a long task will never let the
other tasks to be executed. This is the case for instance if HPO
generation task is a single one that is interrupted after creation of
new configurations.