Handle Interruption after output update
Why:

If the process is interrupted right after the output update, the task
status will be set as failed-over and the report output will be empty,
but the task output will be set. This leads to broken tasks as the next
execution will fail to set the output, which already exists.

How:

Maintainer detects tasks with output and turn them to Completed instead
or queuing them or failing them over.