Add host attrite to tasks
Why:

Tasks may be run on different clusters and may checkpoint themselves
using some local files. Because of that they may only be resumed on the
same cluster. It is also usefull to have information about the node
where the task was executed if there was some error due to system or
hardware failures on the node.

How:

When the task is reserved, info about the host is gathered and saved in
task.report.facility.host. This is updated every time the task is
resumed since the node may be different even thought the cluster is the
same. The attribute is saved as an event sourced item just like the
status so that we can keep a trace of all hosts on all resumptions.

Workers now query tasks that are new (no host) or were previously run on
the same cluster as the worker is currently on.