Make task.resources configurable
Why:

The user can define general resources for an operator, but it is also
possible to overwrite those values in a given task. This is necessary if
the resources requirements depends on the arguments of the operator.

Furthermore, the coming dispatcher will require usage hints in the
resources of task for efficient dispatching of tasks among workers.