Break down worker in Manager/Dispatcher/Worker(s)
Why:

Using the scheduler to assign workers to resources is not fine grain
enough because GPUs can handle many processes but schedulers can only
assign a whole GPU to a job.

How:

A Manager takes care of keeping up 1 dispatcher and 5 workers (should be
configurable). The dispatcher fetches tasks and pile them in a queue
until it assumes the available resources could not handle more. The
usage of resources is estimated based on user hints saved in
task.resources[usage]. In any case, metrics of each task resources usage
is compiled at every heartbeat, so the dispatcher will start using this
data instead when there is at least two snapshot available, which is
both more precise and more hardware/context specific than generic user
hints. Workers pull tasks from the queue and report completion to
dispatcher by piling the task id in the completion queue. If a worker or
the dispatcher unexpectedly die, the Manager will reload them (thanks to
cotyledon).