Reduce fetches for attr updates
Why:

There was a huge amounts of pulls for status and stdout/stderr
updates. In most cases, we only need the last event to create the next
id.

How:

The attributes only fetches the last item when possible. A attribute
`last_item` keeps track of the last item so there is no need to look at
`history` anymore.