Fix attr.refresh by using sort(inc_id)
Why:

There is sometimes swapping of inc_id and timestamp when two events are
logged quickly one after the other from the same application...

Note:

This may actually caused by a bug in mahler. In seams unlikely that
mongodb logs the documents in a different order than the requests order.
It requires further investigations.