Move reservation to dispatcher
Why:

The reservation was kept in workers to make sure that only one worker
could reserve a given task at a time. This makes dispatching unstable
because many dispatcher ends up with the same trials and workers most of
the time fails to reserve and most wait for another trial.

How:

Dispatcher reserve trials before queuing them. It is then responsible to
renew the reservation every `mahler.core.config.heartbeat`. To make sure
that a worker is not running a task that was lost and reserved by
another dispatcher, the reservation is passed a hashcode unique to the
manager and shared to dispatcher and workers. The worker verifies this
code before executing the trial. If the code is different, the worker
fall-backs and wait for another task.