Add flask-caching to limit DB hits
Why:

There was still a huge amount of hits on db, especially for status and
metrics. This was partially due to heartbeat which attempted to refresh
the status every seconds... ouch.

How:

Cache the status and metrics for 60sec when called. Also stop refreshing
the status every seconds, and rather wait for the heartbeat attempt and
catch any changes since then.