Keep track of workers in Dispatcher
Why:

When some workers finish their task, the dispatcher may be sleeping for
`heartbeat` before queuing new tasks. This is wasting a lot of resources
when there is a high turn-over of tasks.

How:

Verify that cached tasks that are not in queue still have status
running, otherwise drop them from the cache. Also, sleep only for 1
second and wait on the completed queue. If something get queued, break
out and fetch again new tasks to queue. If nothing finish after
`heartbeat`, maintain the cache and try updating the queue.